fix: abstract validations for siginscreen to new component authcontent
