+++ App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {SignInScreen} from './src/screens/account/SigInScreen';
+import {SignInScreen} from './src/screens/account/SignInScreen';
 import {NavigationContainer} from '@react-navigation/native';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import {DashboardScreen} from './src/screens/dashboard/DashboardScreen';
@@ -10,7 +10,7 @@ const App = () => {
   return (
     <NavigationContainer>
       <Stack.Navigator>
-        <Stack.Screen name="SigInScreen" component={SignInScreen} />
+        <Stack.Screen name="SignInScreen" component={SignInScreen} />
         <Stack.Screen name="DashboardScreen" component={DashboardScreen} />
       </Stack.Navigator>
     </NavigationContainer>
